Move EthData struct from store to sync package
