--- store/eth_data.go
@@ -1,13 +1,0 @@
-package store
-
-import (
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/openweb3/web3go/types"
-)
-
-type EthData struct {
-	Number   uint64
-	Block    *types.Block
-	Receipts map[common.Hash]*types.Receipt
-	Logs     []types.Log
-}
